Revert "Route legacy dashboard terminal through dashboard resolver"
This reverts commit 3bef5bfa030f57f86e446b5e57dc14497b2e3a4f.
