+++ tests/integration/dashboard-route.test.ts
@@ -21,13 +21,7 @@ import { PROTECTED_URLS } from '@/configs/urls'
  */
 
 // create hoisted mocks
-const {
-  mockAuth,
-  mockCookieStore,
-  mockInfraGet,
-  mockListUserTeams,
-  mockResolveUserTeam,
-} = vi.hoisted(() => ({
+const { mockAuth, mockCookieStore, mockResolveUserTeam } = vi.hoisted(() => ({
   mockAuth: {
     getAuthContext: vi.fn(),
     signOut: vi.fn(),
@@ -36,8 +30,6 @@ const {
     get: vi.fn(),
     set: vi.fn(),
   },
-  mockInfraGet: vi.fn(),
-  mockListUserTeams: vi.fn(),
   mockResolveUserTeam: vi.fn(),
 }))
 
@@ -62,18 +54,6 @@ vi.mock('@/core/server/functions/team/resolve-user-team', () => ({
   resolveUserTeam: mockResolveUserTeam,
 }))
 
-vi.mock('@/core/modules/teams/user-teams-repository.server', () => ({
-  createUserTeamsRepository: vi.fn(() => ({
-    listUserTeams: mockListUserTeams,
-  })),
-}))
-
-vi.mock('@/core/shared/clients/api', () => ({
-  infra: {
-    GET: mockInfraGet,
-  },
-}))
-
 vi.mock('@/lib/utils/cookies', () => ({
   setTeamCookies: vi.fn(),
   getTeamCookies: vi.fn(),
@@ -90,14 +70,6 @@ describe('Dashboard Route - Team Resolution Integration Tests', () => {
       user: { id: 'user-123' },
       accessToken: 'session-token',
     })
-    mockInfraGet.mockResolvedValue({
-      response: { ok: false },
-      data: undefined,
-    })
-    mockListUserTeams.mockResolvedValue({
-      ok: true,
-      data: [],
-    })
   })
 
   afterEach(() => {
@@ -108,10 +80,9 @@ describe('Dashboard Route - Team Resolution Integration Tests', () => {
    * Helper to create a NextRequest with optional query params
    */
   function createRequest(
-    searchParams: Record<string, string> = {},
-    pathname = '/dashboard'
+    searchParams: Record<string, string> = {}
   ): NextRequest {
-    const url = new URL(`http://localhost:3000${pathname}`)
+    const url = new URL('http://localhost:3000/dashboard')
     Object.entries(searchParams).forEach(([key, value]) => {
       url.searchParams.set(key, value)
     })
@@ -158,69 +129,6 @@ describe('Dashboard Route - Team Resolution Integration Tests', () => {
       expect(response.headers.get('location')).toContain(
         '/dashboard/my-team/billing'
       )
-    })
-
-    it('should redirect legacy terminal requests to the team terminal page', async () => {
-      mockResolveUserTeam.mockResolvedValue({
-        id: 'team-456',
-        slug: 'my-team',
-      })
-
-      const request = createRequest(
-        {
-          command: 'ls',
-          template: 'base',
-        },
-        '/dashboard/terminal'
-      )
-
-      const response = await GET(request)
-      const location = response.headers.get('location')
-
-      expect(location).toContain('/dashboard/my-team/terminal')
-      expect(location).toContain('command=ls')
-      expect(location).toContain('template=base')
-      expect(location).not.toContain('__terminal')
-    })
-
-    it('should redirect legacy terminal sandbox requests to the sandbox owner team', async () => {
-      mockResolveUserTeam.mockResolvedValue({
-        id: 'preferred-team',
-        slug: 'preferred',
-      })
-      mockListUserTeams.mockResolvedValue({
-        ok: true,
-        data: [
-          {
-            id: 'owner-team',
-            slug: 'owner',
-          },
-        ],
-      })
-      mockInfraGet
-        .mockResolvedValueOnce({
-          response: { ok: false },
-          data: undefined,
-        })
-        .mockResolvedValueOnce({
-          response: { ok: true },
-          data: { sandboxID: 'ih1km3nxsd8472pml1kkb' },
-        })
-
-      const request = createRequest(
-        {
-          sandboxId: 'ih1km3nxsd8472pml1kkb',
-          template: 'base',
-        },
-        '/dashboard/terminal'
-      )
-
-      const response = await GET(request)
-      const location = response.headers.get('location')
-
-      expect(location).toContain('/dashboard/owner/terminal')
-      expect(location).toContain('sandboxId=ih1km3nxsd8472pml1kkb')
-      expect(location).toContain('template=base')
     })
   })
